refactor(LottoTicket.java):WinnerChecker.java와 병합 refactor(InputView.java):WinnerChecker -> LottoTicket refactor(Lotto.java):Number 인스턴스 변수 제거
