+++ src/main/java/lottoauto/service/LottoTicket.java
@@ -1,22 +1,30 @@
 package lottoauto.service;
 
+import lottoauto.util.Rank;
 import lottoauto.wrapper.Lotto;
-import lottoauto.wrapper.Number;
 import lottoauto.wrapper.Price;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 public class LottoTicket {
+
+
     List<Lotto> listLotto;
     Lotto winnerLotto;
+
     public LottoTicket() {
         this.listLotto = new ArrayList<>();
     }
 
-    public LottoTicket(Lotto winnerLotto) {
-        this.listLotto = new ArrayList<>();
+    public LottoTicket(List<Lotto> listLotto, Lotto winnerLotto) {
+        this.listLotto = listLotto;
         this.winnerLotto = winnerLotto;
+    }
+
+    public List<Lotto> getListLotto() {
+        return listLotto;
     }
 
     public void add(Lotto lotto) {
@@ -54,5 +62,32 @@ public class LottoTicket {
 
     public void getWinnerLotto(Lotto winnerLotto) {
         this.winnerLotto = winnerLotto;
+    }
+
+    public void makeWinnerMap(Map<Integer, Integer> winnerMap) {
+        for (int indexKey = 0; indexKey < listLotto.size(); indexKey++) {
+            addWinnerMapValueByKey(winnerMap, indexKey);
+        }
+    }
+
+    private void addWinnerMapValueByKey(Map<Integer, Integer> winnerMap, int indexKey) {
+        Lotto compareLotto = listLotto.get(indexKey);
+
+        matchedCountCheck(matchedCount(compareLotto, Rank.FIRST), winnerMap, Rank.FIRST);
+        matchedCountCheck(matchedCount(compareLotto, Rank.SECOND) && compareBonusTickets(compareLotto), winnerMap, Rank.SECOND);
+        matchedCountCheck(matchedCount(compareLotto, Rank.THIRD) && !compareBonusTickets(compareLotto), winnerMap, Rank.THIRD);
+        matchedCountCheck(matchedCount(compareLotto, Rank.FOURTH), winnerMap, Rank.FOURTH);
+        matchedCountCheck(matchedCount(compareLotto, Rank.FIFTH), winnerMap, Rank.FIFTH);
+    }
+
+
+    private void matchedCountCheck(boolean compareLotto, Map<Integer, Integer> winnerMap, Rank rank) {
+        if (compareLotto) {
+            winnerMap.put(rank.getLottoRank(), winnerMap.get(rank.getLottoRank()) + 1);
+        }
+    }
+
+    private boolean matchedCount(Lotto compareLotto, Rank rank) {
+        return compareTickets(compareLotto) == rank.getCountOfMatch();
     }
 }
